requirements.txt
Django>=4.2,<5
Pillow>=10.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
